% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_net_surplus.R
\name{calc_net_surplus}
\alias{calc_net_surplus}
\title{This function takes as inputs the unlimited scenario price, supply, and demand, and the
corresponding constrained scenario price, supply, and demand and calculates the change in surplus
The parameter values are obtained from the supply and demand curves outputted by GCAM}
\usage{
calc_net_surplus(cprice, csup, cdmnd, eprice, esup, edmnd)
}
\arguments{
\item{cprice}{the shadow price of water in the unlimited water scenario}

\item{csup}{the supply of water in the unlimited scenario}

\item{cdmnd}{the demand of water in the unlimited scenario}

\item{eprice}{the shadow price of water in the constrained water scenario}

\item{esup}{the supply of water in the constrained water scenario}

\item{edmnd}{the demand of water in the constrained water scenario}
}
\value{
the change in total surplus
}
\description{
This function takes as inputs the unlimited scenario price, supply, and demand, and the
corresponding constrained scenario price, supply, and demand and calculates the change in surplus
The parameter values are obtained from the supply and demand curves outputted by GCAM
}
\author{
Robert Link
}
