% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_query.R
\name{process_query}
\alias{process_query}
\title{This function processes the raw queried data}
\usage{
process_query(dat)
}
\arguments{
\item{dat}{the csv file outputted by the query function}
}
\value{
a data table with all dimensions represented as columns. Levels are numbered according to the tables below


Socioeconomic and agLU dimension numberings correspond to the respective SSP
If other SSP components are linked to the socioeconomic dimension, the level=1, if linked to agriculture, 2
Groundwater Availability: Low groundwater (5%) is coded as 1, medium (25%) as 2, and high  (40%) as 3
Reservoir Storage: Restricted reservoir storage is coded as 1, expanded as 2
Earth System Models: GFDL is 1, HadGEM2 is 2, IPSL is 3, MIROC is 4, NorESM is 5
Land Use Scenario: The FFICT tax is 1, the UCT is 2
}
\description{
This function processes the raw queried data
}
